Chore: Removing unused packages, removing views, updated API routes
