+++ packages/crawler-scanner/package.json
@@ -18,13 +18,10 @@
     "crawler": "^1.3.0",
     "dotenv": "^8.2.0",
     "dotenv-flow": "^3.2.0",
-    "ejs": "~2.6.1",
     "express": "^4.17.1",
-    "http-errors": "~1.6.3",
     "minimist": "^1.2.5",
     "morgan": "~1.9.1",
     "multer": "^1.4.2",
-    "node-sass-middleware": "0.11.0",
     "puppeteer": "^5.5.0",
     "winston": "^3.3.3"
   }
