+++ packages/crawler-scanner/public/index.html
@@ -1,0 +1,10 @@
+<html lang="en">
+<head>
+  <title>Express</title>
+</head>
+
+<body>
+  <h1>Express API server</h1>
+</body>
+
+</html>
