--- packages/crawler-scanner/public/stylesheets/style.css
@@ -1,17 +1,0 @@
-body {
-  padding: 50px;
-  font: 14px "Lucida Grande", Helvetica, Arial, sans-serif; }
-
-a {
-  color: #00B7FF; }
-
-.show {
-  display: block !important; }
-
-.hidden {
-  display: none; }
-
-.green {
-  color: green; }
-
-/*# sourceMappingURL=style.css.map */
