+++ packages/crawler-scanner/src/app.js
@@ -1,61 +1,21 @@
-const createError = require('http-errors');
 const express = require('express');
 const path = require('path');
 const cookieParser = require('cookie-parser');
 const logger = require('morgan');
-const sassMiddleware = require('node-sass-middleware');
 const cors = require('cors');
 const { cwd } = require('process');
-
-const indexRouter = require('./routes/index');
-const scansRouter = require('./routes/scans');
-const launchCrawlerRouter = require('./routes/launch-crawler');
-const launchScannerRouter = require('./routes/launch-scanner');
 
 const apiRouter = require('./routes/api');
 
 const app = express();
-
-// view engine setup
-app.set('views', path.join(__dirname, 'views'));
-app.set('view engine', 'ejs');
 
 app.use(logger('dev'));
 app.use(cors());
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(sassMiddleware({
-  src: path.join(cwd(), 'public'),
-  dest: path.join(cwd(), 'public'),
-  indentedSyntax: true, // true = .sass and false = .scss
-  sourceMap: true
-}));
 app.use(express.static(path.join(cwd(), 'public')));
 
-app.use('/', indexRouter);
-app.use('/scans', scansRouter);
-app.use('/launch-crawler', launchCrawlerRouter);
-app.use('/launch-scanner', launchScannerRouter);
 app.use('/api', apiRouter);
-
-// catch 404 and forward to error handler
-app.use(function (req, res, next) {
-  next(createError(404));
-});
-
-// Ignore the favicon for now
-app.get('/favicon.ico', (req, res) => res.status(204));
-
-// error handler
-app.use(function (err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-  // render the error page
-  res.status(err.status || 500);
-  res.render('error');
-});
 
 module.exports = app;
