+++ packages/crawler-scanner/src/bin/www
@@ -5,7 +5,7 @@
  */
 
 var app = require('../app');
-var debug = require('debug')('express-test-2:server');
+var debug = require('debug')('crawler-scanner:server');
 var http = require('http');
 
 /**
