--- packages/crawler-scanner/src/routes/launch-crawler.js
@@ -1,39 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const path = require('path');
-const childProcess = require('child_process');
-const { cwd } = require('process');
-const winston = require('winston');
-
-const logger = winston.createLogger({
-    transports: [
-        new winston.transports.Console()
-    ]
-});
-
-const libFolderPath = process.env.LIB_FOLDER_PATH;
-
-router.post('/', function (req, res, next) {
-    const a11yLauncherProcess = childProcess.spawn('node', [
-        path.join(cwd(),
-            `${libFolderPath}crawler.js`),
-        '--siteUrl',
-        req.body.siteUrl,
-        req.body.autoScan ? '--autoScan' : '']);
-
-    a11yLauncherProcess.stderr.on('data', (data) => {
-        logger.error(`stderr: ${data}`);
-        res.json({ result: 'error' });
-    });
-
-    a11yLauncherProcess.stdout.on('data', (data) => {
-        logger.info(`stdout: ${data}`);
-    });
-
-    a11yLauncherProcess.on('close', (code) => {
-        logger.info(`crawler exited with code ${code}`);
-        res.json({ result: 'success' });
-    });
-});
-
-module.exports = router;
