--- packages/crawler-scanner/src/views/index.ejs
@@ -1,57 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <title>A11y Crawler POC</title>
-        <meta name="description" content="POC to run the a11y crawler, and display the results in a user-friendly way.">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" href="./stylesheets/tailwind.css">
-        <link rel="stylesheet" href="./stylesheets/style.css">
-    </head>
-    <body>
-        <div class="container">
-            <h1 class="text-3xl">A11Y POC</h1>
-            <h2 class="text-2xl">Available Scan Results</h2>
-            <ul class="mt-3">
-            <% dirs.forEach(dir=>{ %>
-                <li class="leading-6"><a class="text-blue-600" href="./scans/<%=dir %>"><%=dir %></a></li>
-                <%}) %>
-            </ul>
-        </div>
-        
-        <hr class="my-4" />
-        <div class="container flex flex-col">
-            <h2 class="mb-5 text-2xl">Run a New Crawl and Scan</h2>
-            <form class="flex flex-col">
-                <label class="mb-4">Site Url (Without ending slash "/")
-                    <input class="border border-green-500 rounded-sm p-2 block mt-2" name="site-url" id="site-url" type="url" placeholder="https://example.com" pattern="(http|https)://.*" required>
-                </label>
-                <label class="mb-4 flex items-center">After crawl completes, start the scan
-                    <input class="ml-2 border-green-500" type="checkbox" id="auto-scan"/>
-                </label>
-                <input class="cursor-pointer block bg-green-500 p-3 rounded-sm max-w-xs" id="crawl-launch-btn" type="submit" value="Launch Crawler"/>
-            </form>
-            <p id="progress-message" class="mt-3 progress hidden">Your Crawl will now begin. It may take up to 30 minutes. Come back and refresh the page for results.</p>
-            <p id="success-message" class="mt-3 success hidden green">The scan has completed. Please refresh the page to see the results.</p>
-        </div>
-
-        <hr class="my-4" />
-        <div class="container flex flex-col">
-            <h2 class="mb-5 text-2xl">Run a New Scan Only</h2>
-            <form class="flex flex-col" id="scan-form">
-                <label class="mb-4">Upload Your Url Results Text File<br/>(Filename should contain the hostname. IE: www.yahoo.ca.txt)
-                    <input class="border border-green-500 rounded-sm p-2 block mt-2" name="results-file" id="results-file" type="file" accept=".txt" required>
-                </label>
-                <label class="mb-4">File Prefix for Your Scan Files
-                    <input class="border border-green-500 rounded-sm p-2 block mt-2" name="filePrefix" id="filePrefix" type="text" placeholder="example.com" required>
-                </label>
-                <input class="cursor-pointer block bg-green-500 p-3 rounded-sm max-w-xs" id="scan-launch-btn" type="submit" value="Launch Scan"/>
-            </form>
-            <p id="scan-progress-message" class="mt-3 progress hidden">Your Scan will now begin. It may take up to 30 minutes. Come back and refresh the page for results.</p>
-            <p id="scan-success-message" class="mt-3 success hidden green">The scan has completed. Please refresh the page to see the results.</p>
-        </div>
-        
-        <script src="./javascripts/index.js" async defer ></script>
-    </body>
-</html>
