--- packages/crawler-scanner/src/views/scan-details.ejs
@@ -1,50 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <title>A11y Crawler POC</title>
-        <meta name="description" content="POC to run the a11y crawler, and display the results in a user-friendly way.">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" href="/stylesheets/tailwind.css">
-        <link rel="stylesheet" href="/stylesheets/style.css">
-
-        <script>
-          function toggle(id) {
-            document.getElementById(id).classList.toggle('show');
-          }
-        </script>
-    </head>
-    <body>
-        <a class="text-blue-600 mb-3 block" href="./">Go Back</a>
-        <h1 class="text-3xl">A11Y POC</h1>
-        
-        <p class="text-xl">Total violations for <%=results.url %>: <%=results.violations.length %></p>
-
-        <h2 class="text-2xl mb-3">Violation Details</h2>
-        <% results.violations.forEach(function(violation, index){ %>
-            <div class="mt-4 p-4 border border-gray-400 rounded">
-                <h3 class="text-xl mb-3"><%=index+1 %>. <%=violation.description%></h3>
-                <h4 class="text-lg mb-3">Severity: <span style="color:<%=violation.impact == 'minor' ? 'green': 'red'%>;" ><%=violation.impact %></span></h4>
-                <h5 class="text-base mb-3">Tags: 
-                <% violation.tags.forEach(function(tag){ %>
-                  <% if( tag === 'wcag2a') { %>
-                    <a class="text-blue-600" href="https://www.w3.org/WAI/WCAG21/quickref/" target="_blank"><span style="margin-right:10px"><%=tag%></span></a>
-                  <% } else { %>
-                    <span style="margin-right:10px"><%=tag%></span>
-                    <% } %>
-                <% }); %> 
-                </h5>
-                <h5 class="text-base mb-3"><a class="text-blue-600" href="<%=violation.helpUrl %>" target="_blank">More Information About Failure</a></h5>
-                <ol type="a" class="ml-4" style="list-style:lower-alpha !important;">
-                <% violation.nodes.forEach(function(node, idx){ %>
-                    <li class="mb-3"><p class="mb-3"><%=node.failureSummary %></p>
-                    <button class="mb-3 rounded p-2 border-2 border-yellow-500 hover:border-gray-500" id="btn-<%=index%>-<%=idx %>" type="button" onClick="toggle('code-<%=index%>-<%=idx %>')">Show Target Markup</button>
-                    <pre class="bg-blue-100 border border-gray-400 rounded p-3" id="code-<%=index%>-<%=idx %>" style="display:none;white-space:break-spaces;overflow-x:auto;"><%=node.html %></pre>
-                    </li>
-                <% }); %>
-                </ol>
-            </div>
-        <% }); %>
-    </body>
-</html>
