--- packages/crawler-scanner/src/views/scan-summary.ejs
@@ -1,31 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <title>A11y Crawler POC</title>
-        <meta name="description" content="POC to run the a11y crawler, and display the results in a user-friendly way.">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" href="/stylesheets/tailwind.css">
-        <link rel="stylesheet" href="/stylesheets/style.css">
-    </head>
-    <body>
-        <a class="text-blue-600 mb-3 block" href="../">Go Back</a>
-        <h1 class="text-3xl">A11Y POC</h1>
-        <h2 class="text-2xl">Total violations for <%=siteUrl %>: <%=results.totalViolations %></h2>
-        <h3 class="text-xl mt-3">Violations Summary</h3>
-        <ul class="mt-3 mb-3">
-            <li class="leading-6">Minor: <%=results.violationSummary.minor %></li>
-            <li class="leading-6">Moderate: <%=results.violationSummary.moderate %></li>
-            <li class="leading-6">Serious: <%=results.violationSummary.serious %></li>
-            <li class="leading-6">Critical: <%=results.violationSummary.critical %></li>
-        </ul>
-
-        <h3 class="text-xl">Scanned Urls</h3>
-        <ol class="mt-3 list-decimal list-inside">
-            <% results.urlList.forEach(function(url){ %>
-                <li class="leading-6"><a class="text-blue-600" href="./<%=url.resultFile.replace('.json','') %>"><%=url.siteUrl%></a></li>
-                <% }); %> 
-        </ol>
-    </body>
-</html>
