--- packages/crawler-scanner/src/views/scans.ejs
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta charset="utf-8">
-        <meta http-equiv="X-UA-Compatible" content="IE=edge">
-        <title>A11y Crawler POC</title>
-        <meta name="description" content="POC to run the a11y crawler, and display the results in a user-friendly way.">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" href="/stylesheets/tailwind.css">
-        <link rel="stylesheet" href="/stylesheets/style.css">
-    </head>
-    <body>
-        <a class="text-blue-600 mb-3 block" href="../">Go Back</a>
-        <h1 class="text-3xl">A11Y POC</h1>
-        <h2 class="text-2xl">Available Scan Results for <%=siteName %></h2>
-        <ul class="mt-3">
-          <% dirs.forEach(dir=>{ %>
-            <li class="leading-6"><a class="text-blue-600" href="./<%=dir %>"><%=dir %></a></li>
-            <%}) %>
-        </ul>
-    </body>
-</html>
